+++ nfEngine/nfCommonTest/Test.hpp
@@ -9,4 +9,4 @@ const std::string TEST_SAMPLE_FILE_PREFIX = "samplefile_";
 
 // BufferOutputStream and BufferInputStream constants
 const std::string TEXT = "The quick brown fox jumps over the lazy dog";
-const int TEXTSIZE = TEXT.length();
+const size_t TEXTSIZE = TEXT.length();
